requirements.txt
 flask 
 flask-cors 
 flask-migrate 
 flask-jwt-extended 
 flask-sqlalchemy
 flask-bcrypt
 google-genai
 flask-mail
 gunicorn
 python-dotenv
 